Conflicts:
	MMBot.Core/Robot.cs

+++ mmbot/Initializer.cs
@@ -38,7 +38,12 @@ namespace mmbot
             {
                 builder.UseWorkingDirectory(options.WorkingDirectory);
             }
-            
+
+            if (options.Watch)
+            {
+                builder.EnableScriptWatcher();
+            }
+
             Robot robot = null;
 
             try
@@ -90,7 +95,15 @@ namespace mmbot
 
             if (File.Exists(Path.Combine(installationFolder, "mmbot.template.ini")))
             {
-                File.Copy(Path.Combine(installationFolder, "mmbot.template.ini"), Path.Combine(path, "mmbot.ini"));
+                if (File.Exists(Path.Combine(path, "mmbot.ini")))
+                {
+                    Console.WriteLine("mmbot.ini already exists in current directory.");
+                    Console.WriteLine();
+                }
+                else
+                {
+                    File.Copy(Path.Combine(installationFolder, "mmbot.template.ini"), Path.Combine(path, "mmbot.ini"));
+                }
             }
 
             Console.WriteLine("The current directory has been initialized");
